+++ src/app/App.tsx
@@ -1,47 +1,14 @@
-import { useEffect } from "react";
+import { Route, Routes } from "react-router-dom";
 
-import Navbar from "../shared/Navbar";
-import Footer from "../shared/Footer";
-import Hero from "../sections/Hero";
-import About from "../sections/About";
-import Experience from "../sections/Experience";
-import Skills from "../sections/Skills";
-import Contact from "../sections/Contact";
+import HomePage from "../pages/HomePage";
 
 import "../App.css";
 
 function App() {
-  useEffect(() => {
-    // Keep initial hash navigation in sync with active-section tracking.
-    if (window.location.hash) {
-      const element = document.querySelector(window.location.hash);
-
-      if (element) {
-        element.scrollIntoView({
-          behavior: "smooth",
-        });
-      }
-    }
-  }, []);
-
   return (
-    <>
-      <a className="skip-link" href="#main-content">
-        Skip to main content
-      </a>
-
-      <Navbar />
-
-      <main id="main-content" tabIndex={-1}>
-        <Hero />
-        <About />
-        <Experience />
-        <Skills />
-        <Contact />
-      </main>
-
-      <Footer />
-    </>
+    <Routes>
+      <Route path="/" element={<HomePage />} />
+    </Routes>
   );
 }
